updated imports after moving project details
